requirements.txt
requests==2.32.3
urllib3==2.2.3
beautifulsoup4==4.12.3
python-docx==0.8.11
pyyaml==6.0.2
tenacity==8.2.3
rich==13.7.1
